name optimize menu entries the way the model filter names them
availableModels carries target filter values, not model ids, so a grounded
target reached the menu as chatgpt::premium and rendered as its raw value.
Share one label function with the filter dropdown so the two lists agree by
construction.

The web query lookup was keyed by model id for the same reason, and always
missed for a grounded target; the server already filters by the requested
target, so its top query is the answer.
