--- .changeset/gold-pugs-marry.md
@@ -1,5 +1,0 @@
----
-"@workspace/whitelabel": patch
----
-
-Show friendly model names in the optimize menu.
