+++ .changeset/optimize-menu-model-names.md
@@ -1,0 +1,7 @@
+---
+"@workspace/whitelabel": patch
+"@workspace/web": patch
+"@workspace/config": patch
+---
+
+Name the platforms in the optimize menu the way the model filter names them, instead of showing raw ids.
